More big refactors
